+++ client/package.json
@@ -30,6 +30,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.7.1",
+    "react-intersection-observer": "^9.4.1",
     "react-router-dom": "^6.6.2",
     "react-use": "^17.4.0",
     "styled-components": "^5.3.6",
