+++ client/src/components/IllustList.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useMemo, useRef, useState } from "react";
 import styled from "styled-components";
 import ColorTags from "./ColorTags";
-import { blurRange, defaultBoxShadow, getKeyColor } from "@/const/styles";
 import { getImageEndpoint, Tweet, TweetToTag, Work } from "@/utils/api";
 import {
   getImagePixel,
@@ -10,52 +9,10 @@ import {
   DisplayOptions,
   FilterMethod,
 } from "@/utils/utils";
+import IllustGroup from "./IllustGroup";
 
 const Wrapper = styled.div`
   min-height: calc(100vh - 40px);
-  padding: 20px 0;
-  display: flex;
-  gap: 10px;
-  justify-content: space-between;
-  align-items: flex-start;
-`;
-
-const Column = styled.div<{ columnCount: number }>`
-  flex-basis: ${(props) => 100 / props.columnCount}%;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-`;
-
-const Illust = styled.div<{ aspectRatio: number; selected: boolean }>`
-  aspect-ratio: 1 / ${(props) => props.aspectRatio};
-  flex-shrink: 0;
-  flex-grow: 0;
-  border-radius: 2px;
-  box-shadow: ${(props) =>
-    props.selected ? `0 1px 14px ${getKeyColor(0.4)}` : defaultBoxShadow};
-  overflow: hidden;
-  transform: scale(${(props) => (props.selected ? 0.9 : 1.0)});
-  overflow: hidden;
-  transition: box-shadow 0.2s, transform 0.2s;
-
-  &:hover {
-    box-shadow: ${(props) =>
-      props.selected
-        ? `0 1px 16px ${getKeyColor(0.4)})`
-        : "0 1px 4px rgba(0, 0, 0, 0.2)"};
-    transform: scale(${(props) => (props.selected ? 0.89 : 0.98)});
-  }
-`;
-
-const DivImage = styled.div<{ src: string; blur: boolean }>`
-  width: 100%;
-  height: 100%;
-  vertical-align: top;
-  background-image: url("${(props) => props.src}");
-  background-position: center;
-  background-size: cover;
-  filter: blur(${(props) => (props.blur ? blurRange : 0)}px);
 `;
 
 interface IllustListProps {
@@ -93,7 +50,6 @@ const IllustList = ({
 }: IllustListProps) => {
   const [filteredTweets, setFilteredTweets] = useState<Tweet[]>([]);
   const [onlyUnrelatedTweets, setOnlyUnrelatedTweets] = useState<Tweet[]>();
-  const wrapperRef = useRef<HTMLDivElement>(null);
 
   const [pixelColor, setPixelColor] = useState<Color>();
   const [mousePoint, setMousePoint] = useState<[number, number]>();
@@ -153,35 +109,17 @@ const IllustList = ({
     onlyUnrelated,
   ]);
 
-  const getRatio = (tweet: Tweet) =>
-    displayOptions.isSquare
-      ? 1
-      : tweet.Images[0].height / tweet.Images[0].width;
-
-  const tweetRows = useMemo(() => {
-    const totalHeight = filteredTweets
-      .filter((tweet) => tweet.Images.length > 0)
-      .map((tweet) => getRatio(tweet))
-      .reduce((previous, current) => previous + current, 0);
-
-    let tweetsPerColumn: Tweet[][] = [[]];
-    let columnHeight = 0;
-
-    for (const tweet of filteredTweets) {
-      tweetsPerColumn[tweetsPerColumn.length - 1]!.push(tweet);
-      columnHeight += getRatio(tweet);
-      if (
-        columnHeight > totalHeight / displayOptions.columns &&
-        tweetsPerColumn.length < displayOptions.columns
-      ) {
-        tweetsPerColumn.push([]);
-        columnHeight = 0;
-      }
+  const splitedTweets = useMemo(() => {
+    const array: Tweet[][] = [];
+    const n = 100;
+    for (let i = 0; i < Math.ceil(filteredTweets.length / n); i++) {
+      array.push(filteredTweets.slice(i * n, (i + 1) * n));
     }
-    return tweetsPerColumn;
-  }, [filteredTweets, displayOptions]);
+    console.log(array);
+    return array;
+  }, [filteredTweets]);
 
-  const onClickTweet = async (e: React.MouseEvent, id: string) => {
+  /*const onClickTweet = async (e: React.MouseEvent, id: string) => {
     if (e.ctrlKey || e.metaKey) {
       const rect = e.currentTarget.getBoundingClientRect();
 
@@ -227,29 +165,18 @@ const IllustList = ({
       setPixelColor(undefined);
       setMousePoint(undefined);
     }
-  };
+  };*/
 
   return (
     <>
-      <Wrapper ref={wrapperRef}>
-        {tweetRows.map((row, i) => (
-          <Column columnCount={displayOptions.columns} key={i}>
-            {row.map((tweet) => {
-              return (
-                <Illust
-                  aspectRatio={displayOptions.isSquare ? 1 : getRatio(tweet)}
-                  selected={selectedTweetIds.includes(tweet.id)}
-                  onClick={(e) => onClickTweet(e, tweet.id)}
-                  key={tweet.id}
-                >
-                  <DivImage
-                    src={getImageEndpoint(tweet.id, 0)}
-                    blur={process.env.BLUR === "true"}
-                  />
-                </Illust>
-              );
-            })}
-          </Column>
+      <Wrapper>
+        {splitedTweets.map((tweets, index) => (
+          <IllustGroup
+            selectedTweetIds={selectedTweetIds}
+            tweets={tweets}
+            displayOptions={displayOptions}
+            key={index}
+          />
         ))}
       </Wrapper>
       {mousePoint && pixelColor ? (
